some housekeeping to push back the API version bitrot, add go.mod
When listing images with size output we can get what we need from a
HEAD request. GET will retrieve the entire layer/blob so let's not do
that.

Manifest schema version 2 is now in use so made some tweaks to make
use of it.

Since this command is still useful, added some error handling where it
was skipped.

Also cleaned up some debugging in the get() routine